Geometry Nodes Practice
Watching some tutorials around Blender 3D on YouTube while thinking about problem generating random skyscraper sizes, trying to figure a way to create a skyscraper generator for city landscapes.  Created something that could be useful for a Sound Amplitude Viewer, animation is a bit fast to be useful.  For the skyscraper generator I need some way to generate a unique id and use as seed for random values when the object is copied and not sure of a way to do that.