Sirius Flower Garden piece with Emissive Light and Compositing
Reminder it is Black History Month and reduced bias in the World does not amplify ideally.  Art is partially responsible for that and thus offsets matter.  Also important to remember Artists can't stop working for 1 month out of the year does not equate to more stage time is less valuable. https://techinfodebug.wpcomstaging.com/2024/02/21/race-and-love/