4.2 Potential bug on Video Editing
When I use the Video Editor to import a sequence of images it allows me to render the animation using the video editor. When I go back to normal layout, it still seems to use the video editor output instead of the original animation. Might be a way to set source to 3D Scene or Video Editor Animation output. Might be missing something.